+++ internal/app/wwctl/container/exec/child/main.go
@@ -34,30 +34,39 @@ func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 		os.Exit(1)
 	}
 	conf := warewulfconf.Get()
+	runDir := container.RunDir(containerName)
+	if _, err := os.Stat(runDir); os.IsNotExist(err) {
+		return errors.Wrap(err, "container run directory does not exist")
+	}
 	mountPts := conf.MountsContainer
 	mountPts = append(container.InitMountPnts(binds), mountPts...)
 	// check for valid mount points
 	lowerObjects := checkMountPoints(containerName, mountPts)
-	overlayDir := conf.Paths.WWChrootdir + "/overlays"
 	// need to create a overlay, where the lower layer contains
 	// the missing mount points
-	wwlog.Verbose("for ephermal mount use tempdir %s", overlayDir)
-	// ignore errors as we are doomed if a tmp dir couldn't be written
-	_ = os.MkdirAll(path.Join(overlayDir, "work"), os.ModePerm)
-	_ = os.MkdirAll(path.Join(overlayDir, "lower"), os.ModePerm)
-	_ = os.MkdirAll(path.Join(overlayDir, "nodeoverlay"), os.ModePerm)
+	wwlog.Verbose("for ephermal mount use tempdir %s", runDir)
+	if err = os.Mkdir(path.Join(runDir, "work"), os.ModePerm); err != nil {
+		return err
+	}
+	if err = os.Mkdir(path.Join(runDir, "lower"), os.ModePerm); err != nil {
+		return err
+	}
+	if err = os.Mkdir(path.Join(runDir, "nodeoverlay"), os.ModePerm); err != nil {
+		return err
+	}
+	// handle all lower object, have some extra logic if the object is a file
 	for _, obj := range lowerObjects {
 		newFile := ""
 		if !strings.HasSuffix(obj, "/") {
 			newFile = filepath.Base(obj)
 			obj = filepath.Dir(obj)
 		}
-		err = os.MkdirAll(filepath.Join(overlayDir, "lower", obj), os.ModePerm)
+		err = os.Mkdir(filepath.Join(runDir, "lower", obj), os.ModePerm)
 		if err != nil {
 			wwlog.Warn("couldn't create directory for mounts: %s", err)
 		}
 		if newFile != "" {
-			desc, err := os.Create(filepath.Join(overlayDir, "lower", obj, newFile))
+			desc, err := os.Create(filepath.Join(runDir, "lower", obj, newFile))
 			if err != nil {
 				wwlog.Warn("couldn't create directory for mounts: %s", err)
 			}
@@ -65,6 +74,7 @@ func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 		}
 	}
 	containerPath := container.RootFsDir(containerName)
+	// running in a private PID space, so also make / private, so that nothing gets out from here
 	err = syscall.Mount("", "/", "", syscall.MS_PRIVATE|syscall.MS_REC, "")
 	if err != nil {
 		return errors.Wrap(err, "failed to mount")
@@ -72,7 +82,7 @@ func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 	ps1Str := fmt.Sprintf("[%s] Warewulf> ", containerName)
 	if len(lowerObjects) != 0 && nodename == "" {
 		options := fmt.Sprintf("lowerdir=%s,upperdir=%s,workdir=%s",
-			path.Join(overlayDir, "lower"), containerPath, path.Join(overlayDir, "work"))
+			path.Join(runDir, "lower"), containerPath, path.Join(runDir, "work"))
 		wwlog.Debug("overlay options: %s", options)
 		err = syscall.Mount("overlay", containerPath, "overlay", 0, options)
 		if err != nil {
@@ -97,13 +107,13 @@ func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 		}
 		overlays := nodes[0].SystemOverlay.GetSlice()
 		overlays = append(overlays, nodes[0].RuntimeOverlay.GetSlice()...)
-		err = overlay.BuildOverlayIndir(nodes[0], overlays, path.Join(overlayDir, "nodeoverlay"))
+		err = overlay.BuildOverlayIndir(nodes[0], overlays, path.Join(runDir, "nodeoverlay"))
 		if err != nil {
 			wwlog.Error("Could not build overlay: %s", err)
 			os.Exit(1)
 		}
 		options := fmt.Sprintf("lowerdir=%s:%s:%s",
-			path.Join(overlayDir, "lower"), containerPath, path.Join(overlayDir, "nodeoverlay"))
+			path.Join(runDir, "lower"), containerPath, path.Join(runDir, "nodeoverlay"))
 		wwlog.Debug("overlay options: %s", options)
 		err = syscall.Mount("overlay", containerPath, "overlay", 0, options)
 		if err != nil {
@@ -165,11 +175,8 @@ func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 	os.Setenv("PATH", "/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin")
 	os.Setenv("HISTFILE", "/dev/null")
 
-	_ = syscall.Exec(args[1], args[1:], os.Environ())
-	/*
-		Exec replaces the actual program, so nothing to do here afterwards
-	*/
-	return nil
+	wwlog.Debug("Exec: %s %s", args[1], args[1:])
+	return syscall.Exec(args[1], args[1:], os.Environ())
 }
 
 /*