+++ CHANGELOG.md
@@ -60,6 +60,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix the issue that warewulf.conf parse does not support CIDR format. #1130
 - Reduce the number of times syncuser walks the container file system. #1209
 - Create ssh key also when calling `wwctl configure --all` #1250
+- Remove the temporary overlay dir. #1180
+- Remove the temporary overlayfs dir and create them besides rootfs #1180
 
 ### Security
 
@@ -70,6 +72,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Allow specification of the ssh-keys to be to be created. #1185
+
+### Changed
+
+- The command `wwctl container exec` locks now this container during execution. #830
 
 ### Fixed